+++ raiden/tests/integration/long_running/test_stress.py
@@ -20,7 +20,7 @@ from raiden.tests.integration.api.utils import wait_for_listening_port
 from raiden.tests.utils.transfer import assert_synced_channel_state, wait_assert
 from raiden.transfer import views
 from raiden.utils.cli import LogLevelConfigType
-from transport.udp.udp_transport import UDPTransport
+from transport.udp.transport import UDPTransport
 
 log = structlog.get_logger(__name__)
