+++ raiden/tests/unit/test_notifyingqueue.py
@@ -2,7 +2,7 @@ import gevent
 from gevent.event import Event
 
 from raiden.utils.notifying_queue import NotifyingQueue
-from transport.udp.udp_utils import event_first_of
+from transport.udp.utils import event_first_of
 
 
 def add_element_to_queue(queue, element):
