+++ raiden/tests/unit/test_udp_transport.py
@@ -9,7 +9,7 @@ from raiden.network.throttle import TokenBucket
 from raiden.tests.utils.factories import ADDR, UNIT_SECRETHASH, make_address
 from raiden.tests.utils.mocks import MockRaidenService
 from raiden.tests.utils.transport import MockDiscovery
-from transport.udp.udp_transport import UDPTransport
+from transport.udp.transport import UDPTransport
 
 pytestmark = pytest.mark.usefixtures("skip_if_not_udp")
